+++ CHANGELOG.md
@@ -3,6 +3,8 @@ UNRELEASED
 
 * [ [#1369](https://github.com/digitalfabrik/integreat-cms/issues/1369) ] Add contenthash to CSS files for correct cache handling
 * [ [#1046](https://github.com/digitalfabrik/integreat-cms/issues/1046) ] Show number of selected items in lists and page tree
+* [ [#1000](https://github.com/digitalfabrik/integreat-cms/issues/1000) ] Automatically derive location coordinates from address
+* [ [#1180](https://github.com/digitalfabrik/integreat-cms/issues/1180) ] Make coordinates optional for locations not on map
 
 
 2022.4.2